set up the admin view on a separate branch dashboard
