+++ src/App.jsx
@@ -1,20 +1,18 @@
 import { useState } from 'react'
-import Navigation from './components/Navigation'
-import Footer from './components/Footer';
+
 
 import './App.css';
 import './media.css';
 import './admin.css'
-import Hero from './components/Hero';
+import Dashbaord from './pages';
+
 
 function App() {
   
 
   return (
     <div>
-      <Navigation />
-      <Hero />
-      <Footer />
+      <Dashbaord />
     </div>
   )
 }
