--- src/components/Button.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import { useState } from 'react'
-
-export default function Button() {
-    const [btnValue, setbtnValue] = useState("Button Value");
-  return (
-    <button>{btnValue}</button>
-  )
-}
