--- src/components/functions.js
@@ -1,5 +1,0 @@
-const redirectEx = () =>{
-    window.location.href = 'https://play.google.com/store/apps/dev?id=7360729639557371586'
-   } 
-
-   export default redirectEx;
