+++ src/main.jsx
@@ -7,90 +7,20 @@ import {
 } from 'react-router-dom';
 
 import App from './App.jsx';
-import Services from './pages/Services.jsx';
-import Faqs from './pages/Faqs.jsx';
-import Blog from './pages/Blog.jsx';
-import Contact from './pages/Contact.jsx';
-import BlogPost from './pages/blog/BlogPost.jsx';
-import Careeres from './pages/Careeres.jsx';
-import Terms from './pages/Terms.jsx';
-import PrivacyPolicy from './pages/PrivacyPolicy.jsx';
-import errorElement from './components/errorElement.jsx';
-import SkeletonLoader from './components/SeletonLoader.jsx';
-import Layout from './components/Layout.jsx';
-import Ambasador from './pages/Ambasador.jsx';
-import Dashbaord from './pages/admin/index.jsx';
-import Stores from './pages/admin/stores/Stores.jsx';
-import Users from './pages/admin/users/Users.jsx';
-import Agents from './pages/admin/Agents/Agents.jsx';
-import Transactions from './pages/admin/transactions/Transactions.jsx';
-import Orders from './pages/admin/orders/Orders.jsx';
-import Settings from './pages/admin/settings/Settings.jsx';
-import Support from './pages/admin/support/support.jsx';
-import StoreDetails from './pages/admin/stores/StoreDetails.jsx';
+import Dashbaord from './pages/index.jsx';
+import Stores from './pages/stores/Stores.jsx';
+import Users from './pages/users/Users.jsx';
+import Agents from './pages/Agents/Agents.jsx';
+import Transactions from './pages/transactions/Transactions.jsx';
+import Orders from './pages/orders/Orders.jsx';
+import Settings from './pages/settings/Settings.jsx';
+import Support from './pages/support/support.jsx';
+import StoreDetails from './pages/stores/StoreDetails.jsx';
 
 
 
 const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Layout />, // Wrap all routes with Layout
-    children: [
-      {
-        index: true, // Default route
-        element: <App />, // Home page component
-      },
-      {
-        path: "home",
-        element: <App />, // Duplicate for `/home`
-      },
-      {
-        path: "services",
-        element: <Services />,
-      },
-      {
-        path: "faqs",
-        element: <Faqs />,
-      },
-      {
-        path: "blog",
-        element: <Blog />,
-      },
-      {
-        path: "contact",
-        element: <Contact />,
-      },
-      {
-        path: "blogpost",
-        element: <BlogPost />,
-      },
-      {
-        path: "careeres",
-        element: <Careeres />,
-      },
-      {
-        path: "terms",
-        element: <Terms />,
-      },
-      {
-        path: "privacy-policy",
-        element: <PrivacyPolicy />,
-      },
-      {
-        path: "become-ambassador",
-        element: <Ambasador />,
-      },
-      
-    ],
-  },
-  {
-    path: "errorElement",
-    element: <errorElement />,
-  },
-  {
-    path: "loading",
-    element: <SkeletonLoader />,
-  },
+  
   {
     path: "dashboard",
     element: <Dashbaord />,
