--- src/pages/Faqs.jsx
@@ -1,17 +1,0 @@
-import React from 'react'
-import Navigation from '../components/Navigation'
-import FAQ from '../components/Faq'
-import Footer from '../components/Footer';
-import ScrollToTop from '../ScrollToTop'
-
-export default function Faqs() {
-  return (
-    <div>
-        <Navigation />
-        <ScrollToTop />
-        <h1>Frequently Asked Questions</h1>
-        <FAQ />
-        <Footer />
-    </div>
-  )
-}
