--- src/pages/Terms.jsx
@@ -1,81 +1,0 @@
-import React from 'react'
-import Navigation from '../components/Navigation'
-import Footer from '../components/Footer'
-import ScrollToTop from '../ScrollToTop'
-
-export default function Terms() {
-  return (
-    <div className='container'>
-      
-      <Navigation />
-      <ScrollToTop />
-      <h1>Terms And Conditions</h1>
-      <div className="container-narrow">
-        <h3>Introduction</h3>
-        <p>Welcome to Chop ASAP! Chop ASAP is committed to protecting your privacy and personal information. This Privacy Policy describes the types of information we collect, how we use it, and the measures we take to ensure your data remains secure when using our food delivery platform in Cameroon.</p>
-        <h3>1. Information We Collect</h3>
-        <p>We collect information to provide, improve, and secure our services. The types of data we collect include:</p>
-        <h3>1.1 Personal Information</h3>
-        <ul>
-          <li>For Customers: Name, email address, phone number, delivery address, payment information, and order history.</li>
-          <li>For Restaurant Owners: Name, business name, contact information, business location, menu details, and banking/payment information.</li>
-
-        </ul>
-        <h3>1.2 Usage Data</h3>
-        <p>We may collect data about your interaction with our app, such as pages viewed, clicks, time spent, device type, and IP address.</p>
-        <h3>1.3 Location Data</h3>
-        <p>With your consent, we collect and use location data to provide accurate delivery services and update you on order status in real-time.</p>
-        <h3>1.4 Cookies and Tracking Technologies</h3>
-        <p>Our app may use cookies and other tracking technologies to enhance the user experience and personalize your interaction with our platform.</p>
-        <h3>2. How We Use Your Information</h3>
-        <p>We use your data to:</p>
-        <ul>
-          <li>Facilitate and process your orders.</li>
-          <li>Connect customers with restaurants and delivery partners.</li>
-          <li>Communicate with you regarding orders, promotions, and support requests.</li>
-          <li>Improve our services, app functionality, and user experience.</li>
-          <li>Process payments and ensure transaction security.</li>
-          <li>Personalize offers and promotions based on your preferences.</li>
-        </ul>
-        <h3>3. Data Sharing and Disclosure</h3>
-        <p>We may share your data only in the following circumstances:</p>
-        <ul>
-          <li>With Delivery Partners: To provide accurate and timely delivery services.</li>
-          <li>With Restaurants: To fulfill orders and provide service support.</li>
-          <li>With Third-Party Service Providers: For payment processing, data analysis, and other essential business operations. These providers are required to protect your data and only use it for specific purposes.</li>
-          <li>To Comply with Legal Requirements: We may disclose your data if required by law or to protect the rights, property, or safety of Chop ASAP, our users, or others.</li>
-        </ul>
-        <h3>4. Data Security</h3>
-        <p>We take reasonable precautions to protect your personal data from unauthorized access, alteration, or disclosure. Our security measures include data encryption, regular system monitoring, and restricted access to sensitive information.</p>
-        <h3>5. Data Retention</h3>
-        <p>We retain your information only as long as necessary for the purposes outlined in this policy. Afterward, we delete or anonymize your data, except where legal obligations require otherwise.</p>
-        <h3>6. Your Rights</h3>
-        <p>Depending on your location and applicable laws, you may have certain rights over your personal data, including:</p>
-        <ul>
-          <li>Accessing and updating your information.</li>
-          <li>Requesting deletion or restriction of your data.</li>
-          <li>Objecting to data processing or requesting data portability.</li>
-          <h3>7. Children's Privacy</h3>
-        </ul>
-
-        <p>Chop ASAP is not intended for children under 13, and we do not knowingly collect information from individuals under this age.</p>
-        <h3>8. Changes to this Privacy Policy</h3>
-        <p>We may update this policy periodically. When we make changes, we will update the "Effective Date" and notify you as required by law.</p>
-        <h3>9. Contact Us</h3>
-
-        <p>If you have questions, concerns, or wish to exercise your data rights, please contact us at:</p>
-
-        <h3>Chop ASAP Support Team</h3>
-        <p>Chop ASAP Support Team <a href="mailto:support@chopasap.com">support@chopasap.com</a> </p>
-        <p>Address: Yaounde Cameroon, Tradex Simbock </p>
-
-
-
-
-
-      </div>
-
-      <Footer />
-    </div>
-  )
-}
