+++ src/pages/users/Users.jsx
@@ -1,7 +1,7 @@
 
 import SideBar from '../components/SideBar';
 import TopNav from '../components/TopNav';
-import image from '../../../assets/image.png'
+import image from '../../assets/image.png'
 
 function Users() {
   return (
